requirements.txt
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
pandas==2.2.3
tqdm==4.67.1
matplotlib==3.10.0
tabulate==0.9.0
datasets==3.2.0
transformers==4.47.1
faiss-cpu==1.9.0.post1
open_clip_torch==2.29.0
ultralytics==8.3.50
tensordict==0.6.2
